requirements.txt
# Core data & analysis
pandas>=2.0.0
numpy>=1.25.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# I/O & parsing
requests>=2.31.0
pdfplumber>=0.10.0
python-docx>=0.8.11
odfpy>=1.4.1

# Utilities
sparse-dot-topn>=0.3.0

# NLP & ML (stance detection)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0

scipy
